requirements.txt
asgiref==3.7.2
Django==4.2.3
django-bootstrap-modal-forms==3.0.4
django-formset==1.0.1
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
